[rpmdb] request pk-rpm-db-clean to rebuild db if it exists

+++ tools/repair_rpm_db.sh
@@ -1,5 +1,16 @@
 #!/bin/sh
-pkill store
-rm -f /var/lib/rpm/__db.00*
-rpm --rebuilddb
+if ls /usr/libexec/pk-rpm-db-clean; then
+    echo "Leave request to rebuild rpm db on restart"
+    mkdir -p /var/lib/PackageKit
+    touch /var/lib/PackageKit/scheduled-rebuilddb
+else
+    echo "Manually rebuild db and reboot"
+    pkill store
+    pkill packagekitd
+    sleep 4
+    pkill -9 store
+    pkill -9 packagekitd
+    rm -f /var/lib/rpm/__db.00*
+    rpm --rebuilddb
+fi
